Separate PDS agent construction from authentication
Name the helper createPdsAgent and keep it in its own PDS agent module. It resolves an endpoint and constructs an AtpAgent; callers own authentication.
