+++ src/modules/atproto/infrastructure/services/ATProtoAgentService.ts
@@ -1,4 +1,4 @@
-import { createAppPasswordAgent } from './AppPasswordSessionService';
+import { createPdsAgent } from './PdsAgent';
 import { AtpAgent, Agent } from '@atproto/api';
 import { NodeOAuthClient } from '@atproto/oauth-client-node';
 import { IdResolver } from '@atproto/identity';
@@ -178,7 +178,7 @@ export class ATProtoAgentService implements IAgentService {
       const session = appPasswordSessionResult.value;
       if (session) {
         // Create an Agent with the session
-        const { agent } = await createAppPasswordAgent(did.value);
+        const { agent } = await createPdsAgent(did.value);
 
         // Resume the session
         await agent.resumeSession(session);
@@ -231,7 +231,7 @@ export class ATProtoAgentService implements IAgentService {
           await this.appPasswordSessionService.getSession(serviceAccountDid);
         if (existingSessionResult.isOk()) {
           const session = existingSessionResult.value;
-          const { agent } = await createAppPasswordAgent(session.did);
+          const { agent } = await createPdsAgent(session.did);
           await agent.resumeSession(session);
           return ok(agent);
         }
@@ -253,7 +253,7 @@ export class ATProtoAgentService implements IAgentService {
       }
 
       const session = newSessionResult.value;
-      const { agent } = await createAppPasswordAgent(session.did);
+      const { agent } = await createPdsAgent(session.did);
       await agent.resumeSession(session);
 
       return ok(agent);
